lots of changes

+++ Georganize/HomeForm.vb
@@ -11,28 +11,38 @@ Public Class HomeForm
         ' Query to get event name, address , date , start time from user-location and venue-events/venues table combined
         ' select events.name as EName,venues.address as EAdd,date as EDate,[start-time] as ETime from events,[venue-events],venues where events.eventid=[venue-events].eventid and [venue-events].venueid = venues.venueid and visibility = 'public' union select events.name as EName,address as EAdd,date as EDate,[start-time] as ETime from events,[user-location] where events.eventid = [user-location].eventid and visibility = 'public'
         Sqlcmd.Connection = Sqlcon
-        Dim Sqldr As SqlDataReader
 
-        Sqlcmd.CommandText = "select eventid,name from events where visibility = 'public'"
+    End Sub
 
-        Sqldr = Sqlcmd.ExecuteReader
-        Do While Sqldr.Read
-            Dim item As New ListViewItem
-            item.Text = Sqldr.Item(1).ToString
-            item.Tag = Sqldr.Item(0).ToString ' event id
-            PublicEventsList.Items.Add(item)
+    Private Sub RefreshForm() Handles MyBase.VisibleChanged
+        If Me.Visible = True Then
+            PublicEventsList.Items.Clear()
+            Dim Sqldr As SqlDataReader
+            Sqlcmd.Parameters.AddWithValue("Uage", Val(LoggedInUserAge.ToString))
+            Sqlcmd.CommandText = "select eventid,name,date from events where visibility = 'Public' and age <= @Uage order by events.date asc"
 
-        Loop
-        Sqldr.Close()
+            Sqldr = Sqlcmd.ExecuteReader
+            Do While Sqldr.Read
+                Dim item As New ListViewItem
+                item.Text = Sqldr.Item(1).ToString
+                item.Tag = Sqldr.Item(0).ToString ' event id
+                Dim temp As New DateTime
+                temp = Sqldr(2)
+                item.SubItems.Add(temp.ToString("d"))
+                PublicEventsList.Items.Add(item)
 
-        Sqlcmd.Parameters.AddWithValue("usid", LoggedInUser)
-        Sqlcmd.CommandText = "select name from events, users, invitations where users.userid = @usid and users.userid = invitations.recieverid and invitations.eventid = events.eventid;"
-        Sqldr = Sqlcmd.ExecuteReader()
-        Do While Sqldr.Read()
-            InvitationsBox.Items.Add(Sqldr(0))
-        Loop
-        Sqldr.Close()
-        Sqlcmd.Parameters.Clear()
+            Loop
+            Sqldr.Close()
+
+            Sqlcmd.Parameters.AddWithValue("usid", LoggedInUser)
+            Sqlcmd.CommandText = "select name from events, users, invitations where users.userid = @usid and users.userid = invitations.recieverid and invitations.eventid = events.eventid;"
+            Sqldr = Sqlcmd.ExecuteReader()
+            Do While Sqldr.Read()
+                InvitationsBox.Items.Add(Sqldr(0))
+            Loop
+            Sqldr.Close()
+            Sqlcmd.Parameters.Clear()
+        End If
     End Sub
 
     Private Sub ExitApplication(sender As Object, e As EventArgs) Handles MyBase.Closed
@@ -80,13 +90,26 @@ Public Class HomeForm
                 Sqlcmd.Parameters.Clear()
             Else
                 Sqlcmd.Parameters.Clear()
-                Sqlcmd.Parameters.AddWithValue("event", eventid)
-                Sqlcmd.Parameters.AddWithValue("user", LoggedInUser)
-                Sqlcmd.CommandText = "insert into enrollments (eventid, userid) values (@event, @user);"
-                Sqlcmd.ExecuteNonQuery()
-                MessageBox.Show("Successfully enrolled")
-                CodeBox.Text = ""
-                Sqlcmd.Parameters.Clear()
+                Sqlcmd.Parameters.AddWithValue("evnid", eventid)
+                Sqlcmd.CommandText = "select age from events where eventid=@evnid"
+                If LoggedInUserAge < Sqlcmd.ExecuteScalar Then
+                    MessageBox.Show("You cannot join this event as you're under the age limit", "Too young", MessageBoxButtons.OK, MessageBoxIcon.Error)
+                Else
+                    Sqlcmd.Parameters.Clear()
+                    Sqlcmd.Parameters.AddWithValue("event", eventid)
+                    Sqlcmd.Parameters.AddWithValue("user", LoggedInUser)
+                    Sqlcmd.CommandText = "select eventid from enrollments where eventid=@event and userid=@user"
+                    If Sqlcmd.ExecuteScalar <> Nothing Then
+                        MessageBox.Show("You are already enrolled for this event", "Already Enrolled", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                    Else
+                        Sqlcmd.CommandText = "insert into enrollments (eventid, userid) values (@event, @user);"
+                        Sqlcmd.ExecuteNonQuery()
+                        MessageBox.Show("Successfully enrolled into event", "Enrollment Complete", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                        CodeBox.Text = ""
+                        Sqlcmd.Parameters.Clear()
+                    End If
+
+                End If
             End If
         End If
     End Sub
@@ -145,17 +168,44 @@ Public Class HomeForm
             evid = Convert.ToInt16(Sqlcmd.ExecuteScalar())
             Sqlcmd.Parameters.Clear()
 
-            Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
-            Sqlcmd.Parameters.AddWithValue("eventid", evid)
-            Sqlcmd.CommandText = "insert into enrollments (eventid, userid) values (@eventid, @userid);"
-            Sqlcmd.ExecuteNonQuery()
-            Sqlcmd.Parameters.Clear()
+            Sqlcmd.Parameters.AddWithValue("eid", evid)
+            Sqlcmd.CommandText = "select age from events where eventid=@eid"
 
-            Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
-            Sqlcmd.Parameters.AddWithValue("eventid", evid)
-            Sqlcmd.CommandText = "delete from invitations where eventid = @eventid and recieverid = @userid;"
-            Sqlcmd.ExecuteNonQuery()
-            Sqlcmd.Parameters.Clear()
+            If LoggedInUserAge < Sqlcmd.ExecuteScalar Then
+                MessageBox.Show("You cannot join the event: " & evname.ToString & " as you're under the age limit", "Too young", MessageBoxButtons.OK, MessageBoxIcon.Error)
+                Sqlcmd.Parameters.Clear()
+                Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
+                Sqlcmd.Parameters.AddWithValue("eventid", evid)
+                Sqlcmd.CommandText = "delete from invitations where eventid = @eventid and recieverid = @userid;"
+                Sqlcmd.ExecuteNonQuery()
+                Sqlcmd.Parameters.Clear()
+            Else
+                Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
+                Sqlcmd.Parameters.AddWithValue("eventid", evid)
+                Sqlcmd.CommandText = "select eventid from enrollments where eventid=@eventid and userid=@userid"
+                If Sqlcmd.ExecuteScalar <> Nothing Then
+                    MessageBox.Show("You are already enrolled for this event", "Already Enrolled", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                    Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
+                    Sqlcmd.Parameters.AddWithValue("eventid", evid)
+                    Sqlcmd.CommandText = "delete from invitations where eventid = @eventid and recieverid = @userid;"
+                    Sqlcmd.ExecuteNonQuery()
+                    Sqlcmd.Parameters.Clear()
+                Else
+                    Sqlcmd.CommandText = "insert into enrollments (eventid, userid) values (@eventid, @userid);"
+
+                    Sqlcmd.ExecuteNonQuery()
+                    Sqlcmd.Parameters.Clear()
+
+                    Sqlcmd.Parameters.AddWithValue("userid", LoggedInUser)
+                    Sqlcmd.Parameters.AddWithValue("eventid", evid)
+                    Sqlcmd.CommandText = "delete from invitations where eventid = @eventid and recieverid = @userid;"
+                    Sqlcmd.ExecuteNonQuery()
+                    Sqlcmd.Parameters.Clear()
+
+                    MessageBox.Show("You have enrolled into the event: " & evname.ToString & " successfully", "Enrollment Complete", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                End If
+            End If
+
         Next
 
         InvitationsBox.Items.Clear()
